backend/inference/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
vllm==0.6.6.post1
transformers==4.47.1
torch==2.5.1
opencv-python-headless==4.10.0.84
Pillow==11.1.0
decord==0.6.0
celery==5.4.0
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
httpx==0.28.1
